+++ src/Oro/Bundle/WebsiteSearchSuggestionBundle/Tests/Unit/Provider/SuggestionProviderTest.php
@@ -10,19 +10,19 @@ use Oro\Bundle\WebsiteSearchSuggestionBundle\Provider\SuggestionProvider;
 use Oro\Bundle\WebsiteSearchSuggestionBundle\Splitter\PhraseSplitter;
 use Oro\Component\Testing\Unit\EntityTrait;
 use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 
-final class SuggestionProviderTest extends \PHPUnit\Framework\TestCase
+final class SuggestionProviderTest extends TestCase
 {
     use EntityTrait;
 
     private SuggestionProvider $suggestionProvider;
 
     private PhraseSplitter&MockObject $phraseSplitter;
-
     private ProductsProvider&MockObject $productsProvider;
-
     private LocalizationHelper&MockObject $localizationHelper;
 
+    #[\Override]
     protected function setUp(): void
     {
         $this->phraseSplitter = $this->createMock(PhraseSplitter::class);
@@ -34,6 +34,7 @@ final class SuggestionProviderTest extends \PHPUnit\Framework\TestCase
             $this->phraseSplitter,
             $this->localizationHelper
         );
+        $this->suggestionProvider->setChunkSize(10000);
     }
 
     /**
@@ -70,18 +71,12 @@ final class SuggestionProviderTest extends \PHPUnit\Framework\TestCase
             ]);
 
         $result = $this->suggestionProvider->getLocalizedSuggestionPhrasesGroupedByProductId(
-            array_column($skuAndNames, 'id')
+            \array_column($skuAndNames, 'id')
         );
 
         self::assertEquals(
-            [
-                1 => [
-                    "sku_phrase" => [1],
-                    "name_phrase" => [1, 2,],
-                    "sku2_phrase" => [2]
-                ]
-            ],
-            $result
+            ['sku_phrase' => [1 => 1], 'name_phrase' => [1 => 1, 2 => 2], 'sku2_phrase' => [2 => 2]],
+            $result->current()
         );
     }